finalize other routes
