+++ apps/web/ui/placeholders/hero.tsx
@@ -1,4 +1,5 @@
 import { Grid } from "@dub/ui";
+import { cn } from "@dub/utils";
 import { PropsWithChildren } from "react";
 
 const HERO_GRADIENT = `radial-gradient(77% 116% at 37% 67%, #EEA5BA, rgba(238, 165, 186, 0) 50%),
@@ -8,9 +9,17 @@ const HERO_GRADIENT = `radial-gradient(77% 116% at 37% 67%, #EEA5BA, rgba(238, 1
   radial-gradient(90% 136% at 52% 100%, #FD3A4E, rgba(253, 58, 78, 0) 50%),
   radial-gradient(102% 143% at 92% 7%, #72FE7D, rgba(114, 254, 125, 0) 50%)`;
 
-export function Hero({ children }: PropsWithChildren) {
+export function Hero({
+  className,
+  children,
+}: PropsWithChildren<{ className?: string }>) {
   return (
-    <div className="relative mx-auto mt-4 w-full max-w-screen-lg overflow-hidden rounded-2xl bg-neutral-50 p-6 text-center sm:p-20 sm:px-0">
+    <div
+      className={cn(
+        "relative mx-auto mt-4 w-full max-w-screen-lg overflow-hidden rounded-2xl bg-neutral-50 p-6 text-center sm:p-20 sm:px-0",
+        className,
+      )}
+    >
       <Grid
         cellSize={80}
         patternOffset={[1, -58]}
